Abstracted formatOrdinal to util function
